pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tool-anywhere-agent"
version = "0.0.1"
description = "A LangGraph-based agent for tool calling capabilities"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Mad Professor"}
]
maintainers = [
    {name = "Mad Professor"}
]
keywords = ["ai", "langchain", "langgraph", "agent", "tools", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
Homepage = "https://github.com/madprofessor42/tool_anywhere_agent"
Repository = "https://github.com/madprofessor42/tool_anywhere_agent"
Documentation = "https://github.com/madprofessor42/tool_anywhere_agent#readme"
Issues = "https://github.com/madprofessor42/tool_anywhere_agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
